Make HLE DSP load binaries too
